requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
cytoolz==0.11.0
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
oauthlib==3.1.1
parsimonious==0.8.1
protobuf==3.17.3
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rlp==2.0.1
six==1.16.0
sqlparse==0.4.1
toolz==0.11.1
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
web3==5.20.1
websockets==9.1
yarl==1.6.3

celery~=5.1.2
redis~=3.5.3